We can sell an item of any price, as long as we have the price.
BRANCH DESCRIPTION
New branch: client first design, using mock objects, contracts

master brach design approach led us to:
a lot of duplication in test names:
  - several items ....
  - one item ...

We can handle that with:
2 dimensions of tests
  - number of items that the customer tries to buy
     - 0 items
     - 1 item
     - 5 items

  - number of items found:
     - all found
     - some found
     - all not found
     - one not found

In the master branch design approach we had
 only one test with emtpty barcode. Why?
 Combinatoric explosion of test possibilities. We get lazy about writing all these test. We don't even identify them all.

In this approach (client-first)
we can write a test list / speciall cases list as a table:

   | found | not found | empty | none
--------------------------------------------------------
 0 |   x   |     x     |   x
--------------------------------------------------------
 1 |       |           |
--------------------------------------------------------
 5 |       |           |
--------------------------------------------------------
 un|   x   |    x      |   x

1 item : found / not found / empty is easy

but 5 items become a problem
a lot of permutations

it would be nicer to test dimensions separately
  - 1-st dimension:
     - did we add the product to the customer's purchase or not
(that's all that matters)

--------------------------
COMMIT DESCRIPTION

what do we want?
we want something to ask the display to display the price
we're not worried about the details how to display the price
so we want a display interface to display a price
Before, we had to check the price format on the display,
Now we want something to display the price,
Something means interface

We immediately introduce Price whole value to avoid primitive obsession

Why 12345 maps to 795 ?
Before, in the traditional testing suite, we introduced a catalog (dict) for mapping values.
Now we are not interested in making design decisions.

A stub (or fake):
  - I don't mind if you call find_price on catalog or not, but if you
  do it with '12345' I will return Price.cents(795) (irrelevant_price)

A stub is when we tell a method to return a hardcoded result that's convenient for the current test.

This seems a lot of code just for one test:
  - bad news is we had to create all these objects at once
  - good news is we have to implement only SaleController
     - all mock are just interfaces
  - we make a little design upfront to make SaleController work
    - we design the interfaces of Catalog and Display
  - a little risky

we didn't have to triangulate the price to implement self.catalog.find_price in onbarcode because the price is irrelevant in this test