We now more gracefully handle scanning an empty barcode.
empty barcode - ''
this value is significant
in previous tests, barcode or price were irrelevant
such values should be explicitly mark as irrelevant
 - by extracting a constant with apropriate name

We don't need catalog to handle empty barcode sa we pass None.
We could pass non None Catolog and assert catalog.mock_calls == [].
But it seems an overkill and not a good idea.
We don;t have a good enough reason to restrain the test not to touch
catalog.
Better is signaling that we don't need catalog at all. And that's a
signal that this behavior (checking the empty barcode) maybe doesn't
belong here (in the sale controller).