Removed unnecessary constraints from a test.
Our test for 'product found' now checks the behavior more closely.

We remove the possibility to accidentally implement  InMemoryCatalog that returns the only product from the storage, and therefore pass the test without fulfilling the contract

Generally speaking, we want our tests to make the fewest assumptions possible about what they need to run and what want to check. This constrains our code as little as possible and stays out of our way.

We like contracts that are weaker - easier to fulfil

Conclusion to implementing Catalog:

contract semantics became the test list of the implementation
then we found the way to extract contract test from the implementation test

it is possible to do the way around:
 - write contract test first
 - you can't run such test, you can only write it
 - until it is instantiated

What we did so far:
 - we checked that SaleController works (using mocks)
 - we checked that InMemoryCatalog fulfil Catalog's contract

Now we can mark test list of Catalog as DONE