Separated the two steps involved in displaying output to the console.
Now ConsoleDisplay provides a convenient facade of the little
reusable library we created, to makes the API easier to use
  - display function which takes message_template and placeholder
    values (will be deleted within a few next commits)
  - render and format functions

API:
  - display_product_not_found_message
  - display_empty_barcode_message
  - display_price
The client of the API takes care of matching arguments -
message_template and placeholder_values.

Displaying something to the user usually consists of 2 behaviors:
 - formatting the output
 - rendering the output

What we have here is a tiny little library

ConsoleDisplay is a mediator/adaptor that coordinates:
 - text formatting engine which calls str.format
 - text rendering engine which calls console print

These 2 behaviors render and format have nothing to do with eachother.
They don't use any fields so they can move to a separate class/module.