ConsoleDisplay now implements display.
We are sure ConsoleDisplay respects the syntacs of the Catalog
contract.
With semantics  - we are not sure because there are no any
details about it in tests. It means that the SaleController doesn't
care what exactly Display does whithin those methods.

In other words:
as long as ConsoleDisplay claims that it can
display_product_not_found_message, empty_barcode_message and price,
and accept those parameters, the Display contract is fulfilled.

There are no contract tests to write for
ConsoleDisplay,
because there is only syntacs contract, and it is checked
by python (that all Display methods have been implemented)

In other languages it would have to be checked by hand.

This contract is vwry weak - the application of the display can do
what it wants.