Made production code available to the first generation of our app.
We don't want production code to depend on tests.
So we have to pull / copy sale.py to production catalog.

It may end up copying stuff into production package a few times as it
turns out that we need new things. It's better because we are sure we
don;t move unnecessary thing into production. We should as little code
into production as possible - only neccessary parts.