We can interpret exactly zero command.
Abstarcting away input reader from stdin

Command line suggests stdin, but we are abstracting it away
Thus the name 'interpret text commands'
It doesn't care where are commands from

We are mocking a class which already exists:
 - SellOneItemController
 - we don't have interface for it
 - all we know is that there is onbarcode method
 - if we don;t know the intention revealing name
 - let's give it a structural/mechanical name
 - if onbarcode is an event then the object is BarcodeScannedListener
 - passed tense

How should interpreter handle empty lines ?
Remember the smell from controller.
Now when we design the object which stands in front of the controller,
we can examin should handle the empty barcode.